Delete source /.DS_Store
According to Nate ;Because it does not currently match, several automatically-used files (such as, most importantly, the .DS_Store files in each folder) that belong to the MacOS operating system are being tracked by Github Desktop every time those files are refreshed by a computer. That makes it hard to synchronize the repository, because the .DS_Store files are refreshed by MacOS on the local computer and they conflict with the duplicate files stored in the GitHub repository.